+++ src/Interfaces/NetworkADT.java
@@ -20,7 +20,7 @@ public interface NetworkADT<T> extends GraphADT<T> {
      * @param vertex2 the second vertex
      * @param weight the weight
      */
-    public void addEdge(T vertex1, T vertex2, int weight);
+    public void addEdge(T vertex1, T vertex2, double weight);
 
     /**
      * Returns the weight of the shortest path in this network
